fix: refactor이후 까먹고 변경하지 않은 메서드 수정
